DTVRecorder: Fix 32 to 64 bit integer arithmetic in FindPSKeyFrames()
This is reported to fix commercial flagging of PVR-150/300/500 recordings
possibly only affecting 32 bit O/Ss due to a bad conversion from a 32 bit
unsigned int to a 64 bit signed int.

Fixes #12313.

+++ mythtv/libs/libmythtv/recorders/dtvrecorder.cpp
@@ -1173,7 +1173,7 @@ void DTVRecorder::FindPSKeyFrames(const uint8_t *buffer, uint len)
         if (hasKeyFrame)
         {
             _last_keyframe_seen = _frames_seen_count;
-            HandleKeyframe(_payload_buffer.size() - (bufptr - bufstart));
+            HandleKeyframe((int64_t)_payload_buffer.size() - (bufptr - bufstart));
         }
 
         if ((aspectRatio > 0) && (aspectRatio != m_videoAspect))
